Added Saves and CSV download from FTP

+++ Budget/app/src/main/java/com/thinus/budget/Database.java
@@ -1,5 +1,6 @@
 package com.thinus.budget;
 
+import android.app.ProgressDialog;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
@@ -30,7 +31,7 @@ public class Database {
         db = tmpdb;
 
         db.execSQL("CREATE TABLE IF NOT EXISTS [category] (id INT PRIMARY KEY NOT NULL, name VARCHAR, categorytype INT, budget DOUBLE, smsDescription VARCHAR);");
-        db.execSQL("CREATE TABLE IF NOT EXISTS [transaction] (id INT PRIMARY KEY NOT NULL, amount DOUBLE, categoryid INT, account VARCHAR,description VARCHAR,date VARCHAR, reference VARCHAR, income_expense int);");
+        db.execSQL("CREATE TABLE IF NOT EXISTS [transaction] (id INT PRIMARY KEY NOT NULL, amount DOUBLE, categoryid INT, account VARCHAR,description VARCHAR,date VARCHAR, reference VARCHAR, income_expense int, deleted int, sms VARCHAR, csv_prov VARCHAR, csv_hist VARCHAR, date_created VARCHAR, split int, split_trans_id1 int, split_trans_id2 int);");
         db.execSQL("CREATE TABLE IF NOT EXISTS [settings] (id INT PRIMARY KEY NOT NULL, name VARCHAR,value VARCHAR);");
     }
 
@@ -151,10 +152,26 @@ public class Database {
      */
 
     public static void LoadTransactionsFromDB() {
-        LoadTransactionsFromDB("Date DESC");
+        LoadTransactionsFromDB("Date DESC", 0);
+    }
+    public static void LoadTransactionsFromDB(int type) {
+        LoadTransactionsFromDB("Date DESC", type);
     }
     public static void LoadTransactionsFromDB(String sort) {
-        Cursor resultSet = db.rawQuery("SELECT * FROM [transaction] ORDER BY " + sort, null);
+        LoadTransactionsFromDB(sort, 0);
+    }
+    public static void LoadTransactionsFromDB(String sort, int type) {
+        Cursor resultSet = null;
+
+
+        if (type == 0) {
+            resultSet= db.rawQuery("SELECT * FROM [transaction] ORDER BY " + sort, null);
+        } else {
+            if (type == 1) {
+                resultSet = db.rawQuery("SELECT * FROM [transaction_possible] ORDER BY " + sort, null);
+            }
+        }
+
         resultSet.moveToFirst();
 
         while (!resultSet.isAfterLast()) {
@@ -174,32 +191,78 @@ public class Database {
             }
 
             //Date date = new Date(resultSet.getLong(5)*1000);
-            String reference = resultSet.getString(6);
-            if (reference.contains("//")) {
-                reference = reference.replace("//","'");
-            }
+            String reference = "";
+            //String reference = resultSet.getString(6);
+            //if (reference.contains("//")) {
+            //    reference = reference.replace("//","'");
+            //}
             int income_expence = resultSet.getInt(7);
+            int deleted = resultSet.getInt(8);
+            String sms = resultSet.getString(9);
+            if (sms == null)
+                sms = "";
+            if (sms.contains("//")) {
+                sms = sms.replace("//","'");
+            }
+            String csv_prov = resultSet.getString(10);
+            if (csv_prov == null)
+                csv_prov = "";
+            if (csv_prov.contains("//")) {
+                csv_prov = csv_prov.replace("//","'");
+            }
+            String csv_hist = resultSet.getString(11);
+            if (csv_hist == null)
+                csv_hist = "";
+            if (csv_hist.contains("//")) {
+                csv_hist = csv_hist.replace("//","'");
+            }
+            Date date_created = null;
+            try {
+                date_created = MainActivity.dateFormat.parse(resultSet.getString(12));
+            } catch (ParseException e) {
+                e.printStackTrace();
+            }
+            int split = resultSet.getInt(13);
+            int split_trans_id1 = resultSet.getInt(14);
+            int split_trans_id2 = resultSet.getInt(15);
 
             Transaction sr1 = null;
-            sr1 = new Transaction(id, amount, categoryId, account, description, date, reference, income_expence);
-            MainActivity.transactionItems.add(sr1);
+            sr1 = new Transaction(id, amount, categoryId, account, description, date, reference, income_expence, deleted, sms, csv_prov, csv_hist, date_created, split, split_trans_id1, split_trans_id2);
+            sr1.setType(type);
+            if (type == 0) {
+                MainActivity.transactionItems.add(sr1);
+            } else {
+                if (type == 1) {
+                    MainActivity.transactionItemsPossibleLink.add(sr1);
+                }
+            }
             resultSet.moveToNext();
         }
     }
 
+    public static void SaveTransaction(Transaction tr) {
+        String table = "transaction";
+        if (tr.getType()==1) {
+            table = "transaction_possible";
+        }
+        db.execSQL("INSERT INTO [" + table + "] VALUES('" + tr.getId() + "','" +
+                tr.getAmount() + "','" +
+                tr.getCategoryId() + "','" +
+                tr.getAccount() + "','" +
+                tr.getDescription() + "','" +
+                tr.getDateFormated() + "','" +
+                //tr.getReference() + "','" +
+                "" + "','" +
+                tr.getIncomeExpense() + "','" +
+                tr.getDeleted() + "','" +
+                tr.getSms() + "','" +
+                tr.getCSV_prov() + "','" +
+                tr.getCSV_hist() + "','" +
+                tr.getDateCreated() + "','" +
+                tr.getSplit() + "','" +
+                tr.getSplitTransID1() + "','" +
+                tr.getSplitTransID2() + "');");
 
-
-
-
-    public static void SaveTransaction(Transaction tr){
-        db.execSQL("INSERT INTO [transaction] VALUES('" +   tr.getId() + "','" +
-                                                            tr.getAmount() + "','" +
-                                                            tr.getCategoryId() + "','" +
-                                                            tr.getAccount() + "','" +
-                                                            tr.getDescription() + "','" +
-                                                            tr.getDateFormated() + "','" +
-                                                            tr.getReference() + "','" +
-                                                            tr.getIncomeExpense() + "');");
     }
 
     public static void UpdateTransaction(Transaction tr) {
@@ -208,16 +271,33 @@ public class Database {
                                             "', account = '" + tr.getAccount() +
                                             "', description = '" + tr.getDescription() +
                                             "', date = '" + tr.getDateFormated() +
-                                            "', reference = '" + tr.getReference() +
+                                            //"', reference = '" + tr.getReference() +
+                                            "', reference = '" + "" +
                                             "', income_expense = '" + tr.getIncomeExpense() +
+                                            "', deleted = '" + tr.getDeleted() +
+                                            "', sms = '" + tr.getSms() +
+                                            "', csv_prov = '" + tr.getCSV_prov() +
+                                            "', csv_hist = '" + tr.getCSV_hist() +
+                                            "', date_created = '" + tr.getDateCreated() +
+                                            "', split = '" + tr.getSplit() +
+                                            "', split_trans_id1 = '" + tr.getSplitTransID1() +
+                                            "', split_trans_id2 = '" + tr.getSplitTransID2() +
                 "' WHERE id = " + tr.getId() + ";");
     }
 
     public static void DeleteTransaction(Transaction tr) {
-        db.execSQL("DELETE FROM [transaction] WHERE id = " + tr.getId() + ";");
+
+        if (tr.getType()==1) {
+            db.execSQL("DELETE FROM [transaction_possible] WHERE id = " + tr.getId() + ";");
+        } else {
+            if (tr.getDeleted() == 0) {
+                tr.setDeleted(1);
+                db.execSQL("UPDATE [transaction] SET [deleted] = '" + tr.getDeleted() + "' WHERE id = " + tr.getId() + ";");
+            } else {
+                db.execSQL("DELETE FROM [transaction] WHERE id = " + tr.getId() + ";");
+            }
+        }
     }
-
-
 
     /*
     Categories
@@ -294,6 +374,37 @@ public class Database {
             }
 
             return "DB Fixed";
+        } catch (Exception ex) {
+            message = ex.toString();
+        }
+        return message;
+    }
+
+    public static String UpgradeDB() {
+        String message = null;
+        try
+        {
+            //db.execSQL("ALTER TABLE [transaction] ADD COLUMN deleted int;");
+            //db.execSQL("ALTER TABLE [transaction] ADD COLUMN sms VARCHAR;");
+            //db.execSQL("ALTER TABLE [transaction] ADD COLUMN csv_prov VARCHAR;");
+            //db.execSQL("ALTER TABLE [transaction] ADD COLUMN csv_hist VARCHAR;");
+            //db.execSQL("ALTER TABLE [transaction] ADD COLUMN date_created VARCHAR;");
+            //db.execSQL("ALTER TABLE [transaction] ADD COLUMN split int;");
+            //db.execSQL("ALTER TABLE [transaction] ADD COLUMN split_trans_id1 int;");
+            //db.execSQL("ALTER TABLE [transaction] ADD COLUMN split_trans_id2 int;");
+
+            //db.execSQL("CREATE TABLE IF NOT EXISTS [transaction_possible] (id INT PRIMARY KEY NOT NULL, amount DOUBLE, categoryid INT, account VARCHAR,description VARCHAR,date VARCHAR, reference VARCHAR, income_expense int, deleted int, sms VARCHAR, csv_prov VARCHAR, csv_hist VARCHAR, date_created VARCHAR, split int, split_trans_id1 int, split_trans_id2 int);");
+            db.execSQL("Delete from [transaction_possible]");
+
+
+            //db.execSQL("UPDATE [transaction] SET [sms] = [reference] WHERE [reference] like 'Standard Bank:%' ");
+            //db.execSQL("UPDATE [transaction] SET [deleted] = 0,[date_created] = '2000-01-01 00:00:00', split = 0, split_trans_id1 = 0, split_trans_id2 = 0 ");
+            //db.execSQL("UPDATE [transaction] SET [csv_prov] = [reference] WHERE [reference] like '\"PROV\",%' ");
+
+            //db.execSQL("UPDATE [transaction] SET [reference] ='' ");
+
+
+            return "DB Upgraded";
         } catch (Exception ex) {
             message = ex.toString();
         }
